src/ethics.py
```python
# Importing necessary libraries
from aif360.algorithms.preprocessing import Reweighing
from aif360.datasets import BinaryLabelDataset
from themis_ml.metrics import mean_difference
from themis_ml.postprocessing import Relabeller
from auditai.fairness import check_algorithmic_fairness

class Ethics:
    def __init__(self, model):
        self.model = model

    def reweighing(self, dataset):
        """
        Function to mitigate bias in the dataset using Reweighing
        """
        rw = Reweighing(unprivileged_groups=[{'race': 1}],
                        privileged_groups=[{'race': 0}])
        dataset_transf = rw.fit_transform(dataset)
        return dataset_transf

    def mean_difference(self, y_true, y_pred):
        """
        Function to calculate the mean difference to check fairness
        """
        mean_diff = mean_difference(y_true, y_pred)
        return mean_diff

    def relabelling(self, y_true, y_pred):
        """
        Function to relabel the predictions to mitigate bias
        """
        relabeller = Relabeller()
        y_pred_relab = relabeller.fit_transform(y_true, y_pred)
        return y_pred_relab

    def check_algorithmic_fairness(self, y_true, y_pred, protected_class):
        """
        Function to check algorithmic fairness of the model
        """
        fairness = check_algorithmic_fairness(y_true, y_pred, protected_class)
        return fairness
```
