tests/test_ethics.py
```python
# Importing necessary libraries
import unittest
from src.ethics import Ethics
from aif360.datasets import BinaryLabelDataset
import numpy as np

class TestEthics(unittest.TestCase):
    def setUp(self):
        """
        Function to set up the necessary objects for testing
        """
        self.model = Ethics(None)
        self.dataset = BinaryLabelDataset(
            favorable_label=1,
            unfavorable_label=0,
            df=pd.DataFrame(np.random.randint(0,2,size=(100, 4)), columns=list('ABCD')),
            label_names=['A'], protected_attribute_names=['B']
        )
        self.y_true = np.random.randint(0,2,size=(100,))
        self.y_pred = np.random.randint(0,2,size=(100,))
        self.protected_class = 'B'

    def test_reweighing(self):
        """
        Function to test the reweighing method
        """
        dataset_transf = self.model.reweighing(self.dataset)
        self.assertIsInstance(dataset_transf, BinaryLabelDataset)

    def test_mean_difference(self):
        """
        Function to test the mean difference method
        """
        mean_diff = self.model.mean_difference(self.y_true, self.y_pred)
        self.assertIsInstance(mean_diff, float)

    def test_relabelling(self):
        """
        Function to test the relabelling method
        """
        y_pred_relab = self.model.relabelling(self.y_true, self.y_pred)
        self.assertIsInstance(y_pred_relab, np.ndarray)

    def test_check_algorithmic_fairness(self):
        """
        Function to test the check algorithmic fairness method
        """
        fairness = self.model.check_algorithmic_fairness(self.y_true, self.y_pred, self.protected_class)
        self.assertIsInstance(fairness, bool)

if __name__ == '__main__':
    unittest.main()
```
